requirements.txt
onnx
onnxruntime-gpu
pillow
numpy

--extra-index-url https://download.pytorch.org/whl/cu128
torch
torchvision